pyproject.toml
[project]
name = "ranger-twin"
version = "0.1.0"
description = "Agent-first AI platform for post-fire forest recovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Internal"}
authors = [
    {name = "TechTrend Federal", email = "team@techtrend.federal"}
]
keywords = ["digital-twin", "forestry", "ai", "geospatial", "usfs"]

dependencies = [
    "google-adk>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | dist
  | build
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["twin_core", "agent_common"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = [
    "venv",
    ".venv",
    "__pycache__",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["agents", "services", "packages"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=agents --cov=services --cov=packages --cov-report=term-missing"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
